requirements.txt
Flask-Smorest
Flask-SQLAlchemy
passlib
marshmallow
gunicorn
flask-migrate
psycopg2-binary